requirements.txt
streamlit==1.46.1
torch==2.7.1
opencv-python-headless==4.12.0.88
numpy==2.2.6
Pillow==11.3.0
torchvision==0.15.2
pytorch-grad-cam==1.5.5
dlib==20.0.0
